started cleaning up tests, fixed fixtures loading error, removed error when contest time limit is uninitialized

+++ db/migrate/020_add_graders_right_to_admin_role.rb
@@ -12,6 +12,6 @@ class AddGradersRightToAdminRole < ActiveRecord::Migration
 
   def self.down
     graders_right = Right.find_by_name('graders_admin')
-    graders_right.destroy
+    graders_right.destroy if graders_right
   end
 end
