+++ db/migrate/025_add_site_to_user_and_add_default_site.rb
@@ -24,6 +24,6 @@ class AddSiteToUserAndAddDefaultSite < ActiveRecord::Migration
     remove_column :users, :site_id
 
     default_site = Site.find_by_name('default')
-    default_site.destroy
+    default_site.destroy if default_site
   end
 end
